+++ sdk/go/core/tests/service_netconf_test.go
@@ -13,6 +13,8 @@ import (
 	"github.com/CiscoDevNet/ydk-go/ydk/services"
 	"github.com/CiscoDevNet/ydk-go/ydk/types"
 	"github.com/CiscoDevNet/ydk-go/ydk/types/datastore"
+	"github.com/CiscoDevNet/ydk-go/ydk/types/yfilter"
+	"github.com/CiscoDevNet/ydk-go/ydk/types/ylist"
 	"github.com/stretchr/testify/suite"
 	"testing"
 )
@@ -314,6 +316,54 @@ func (suite *NetconfServiceTestSuite) TestSanityGetRunningConfig() {
     for _, entity := range getEC.Entities() {
     	ydk.YLogDebug(fmt.Sprintf("Printing %s", GetEntityXMLString(entity)))
     }
+}
+
+func (suite *NetconfServiceTestSuite) TestDeleteContainer() {
+	// Build loopback configuration
+	address := ysanity.Native_Interface_Loopback_Ipv4_Address{}
+	address.Ip = "2.2.2.2"
+	address.PrefixLength = 32
+
+	loopback := ysanity.Native_Interface_Loopback{}
+	loopback.Name = 2222
+	loopback.Ipv4.Address = append(loopback.Ipv4.Address, &address)
+
+	native := ysanity.Native{}
+        native.Interface.Loopback = append(native.Interface.Loopback, &loopback)
+
+	result := suite.NS.EditConfig(
+		&suite.Provider, datastore.Candidate, &native, "", "", "")
+	suite.True(result)
+
+	// Read ipv4 configuration
+	native = ysanity.Native{}
+	loopback = ysanity.Native_Interface_Loopback{}
+	loopback.Name = 2222
+	native.Interface.Loopback = append(native.Interface.Loopback, &loopback)
+	types.SetAllParents(&native)
+	ipv4ConfigEnt := suite.NS.GetConfig(
+		&suite.Provider, datastore.Candidate, &loopback.Ipv4)
+	suite.NotNil(ipv4ConfigEnt)
+	ipv4Config := ipv4ConfigEnt.(*ysanity.Native_Interface_Loopback_Ipv4)
+	_, addressEnt := ylist.Get(ipv4Config.Address, "2.2.2.2")
+	suite.NotNil(addressEnt)
+	addressPtr := addressEnt.(*ysanity.Native_Interface_Loopback_Ipv4_Address)
+	suite.Equal("32", addressPtr.PrefixLength)
+
+	// Remove ipv4 configuration
+	native = ysanity.Native{}
+	loopback = ysanity.Native_Interface_Loopback{}
+	loopback.Name = 2222
+	native.Interface.Loopback = append(native.Interface.Loopback, &loopback)
+	types.SetAllParents(&native)
+	loopback.Ipv4.YFilter = yfilter.Delete
+	result = suite.NS.EditConfig(
+		&suite.Provider, datastore.Candidate, &loopback.Ipv4, "", "", "")
+        suite.True(result)
+
+	// Delete configuration
+	result = suite.NS.DiscardChanges(&suite.Provider)
+	suite.True(result)
 }
 
 func TestNetconfServiceTestSuite(t *testing.T) {