+++ sdk/python/core/ydk/services/netconf_service.py
@@ -16,9 +16,12 @@
 from ydk.ext.services import Datastore, NetconfService as _NetconfService
 from ydk.errors import YServiceError as _YServiceError
 from ydk.errors.error_handler import handle_runtime_error as _handle_error
+from ydk.filters import YFilter
 
 from ydk.types import EntityCollection, Config
 from ydk.entity_utils import _read_entities, _get_top_level_entity, _get_child_entity_from_top
+from ydk.entity_utils import _set_nontop_entity_filter
+
 
 class NetconfService(_NetconfService):
     """ Python wrapper for NetconfService
@@ -88,7 +91,7 @@ class NetconfService(_NetconfService):
             return self._ns.discard_changes(provider)
 
     def edit_config(self, provider, target, config,
-        default_operation="", test_option="", error_option=""):
+                    default_operation="", test_option="", error_option=""):
 
         if None in (provider, target, config):
             raise _YServiceError("provider, target, and config cannot be None")
@@ -97,7 +100,7 @@ class NetconfService(_NetconfService):
             if isinstance(config, Config):
                 config = config.entities()
             return self._ns.edit_config(provider, target, config,
-                default_operation, test_option, error_option)
+                                        default_operation, test_option, error_option)
 
     def get_config(self, provider, source=Datastore.running, read_filter=None):
         if None in (provider, source):
@@ -111,6 +114,7 @@ class NetconfService(_NetconfService):
         if isinstance(read_filter, EntityCollection):
             filters = read_filter.entities()
 
+        _set_nontop_entity_filter(filters, YFilter.read)
         top_filters = _get_top_level_entity(filters, provider.get_session().get_root_schema())
         with _handle_error():
             top_result = self._ns.get_config(provider, source, top_filters)
@@ -132,6 +136,7 @@ class NetconfService(_NetconfService):
         if isinstance(read_filter, EntityCollection):
             filters = read_filter.entities()
 
+        _set_nontop_entity_filter(filters, YFilter.read)
         top_filters = _get_top_level_entity(filters, provider.get_session().get_root_schema())
         with _handle_error():
             top_result = self._ns.get(provider, top_filters)